+++ .github/CODEOWNERS
@@ -21,6 +21,7 @@
 /tracer/src/Datadog.Trace/Iast/           @DataDog/asm-dotnet
 AspectsDefinitions.g.cs 				@DataDog/asm-dotnet
 /tracer/test/test-applications/integrations/Samples.InstrumentedTests/ @DataDog/asm-dotnet
+/tracer/src/Datadog.Trace/Tags.AppSec.cs @DataDog/asm-dotnet
 
 # Profiler
 /profiler/                                @DataDog/profiling-dotnet
