+++ tracer/src/Datadog.Trace/Tags.cs
@@ -8,7 +8,7 @@ namespace Datadog.Trace
     /// <summary>
     /// Standard span tags used by integrations.
     /// </summary>
-    public static class Tags
+    public static partial class Tags
     {
         /// <summary>
         /// The environment of the instrumented service. Its value is usually constant for the lifetime of a process,
@@ -582,55 +582,6 @@ namespace Datadog.Trace
         /// Marks a span as injected when DBM data was propagated
         /// </summary>
         internal const string DbmDataPropagated = "_dd.dbm_trace_injected";
-
-        internal static class AppSec
-        {
-            internal const string Events = "appsec.events.";
-
-            internal static string Track(string eventName) => $"{Events}{eventName}.track";
-
-            internal static class EventsUsers
-            {
-                internal const string EventsUsersRoot = Events + "users";
-
-                internal static class LoginEvent
-                {
-                    private const string Root = EventsUsersRoot + ".login";
-                    internal const string Success = Root + ".success";
-                    internal const string SuccessSdkSource = $"_dd.{Success}.sdk";
-                    internal const string SuccessAutoMode = $"_dd.{Success}.auto.mode";
-                    internal const string SuccessTrack = Success + ".track";
-
-                    internal const string Failure = Root + ".failure";
-                    internal const string FailureUserId = Failure + ".usr.id";
-                    internal const string FailureUserExists = Failure + ".usr.exists";
-                    internal const string FailureEmail = Failure + ".email";
-                    internal const string FailureUserName = Failure + ".username";
-                    internal const string FailureAutoMode = $"_dd.{Failure}.auto.mode";
-                    internal const string FailureSdkSource = $"_dd.{Failure}.sdk";
-                    internal const string FailureTrack = Failure + ".track";
-                }
-
-                internal static class SignUpEvent
-                {
-                    private const string Root = EventsUsersRoot + ".signup";
-                    private const string Success = Root + ".success";
-                    internal const string SuccessUserId = Root + ".usr.id";
-                    internal const string SuccessEmail = Root + ".usr.email";
-                    internal const string SuccessUserName = Root + ".usr.username";
-
-                    internal const string SuccessAutoMode = $"_dd.{Success}.auto.mode";
-                    internal const string SuccessTrack = Success + ".track";
-
-                    private const string RootFailure = Root + ".failure";
-                    internal const string FailureUserId = RootFailure + ".usr.id";
-                    internal const string FailureEmail = RootFailure + ".usr.email";
-                    internal const string FailureUserName = RootFailure + ".usr.username";
-                    internal const string FailureAutoMode = $"_dd.{RootFailure}.auto.mode";
-                    internal const string FailureTrack = RootFailure + ".track";
-                }
-            }
-        }
 
         internal static class User
         {
